+++ library/src/schemas/picklist/types.ts
@@ -1,6 +1,8 @@
+import type { MaybeReadonly } from '../../types/index.ts';
+
 /**
  * Picklist options type.
  */
-export type PicklistOptions<TOption extends string = string> =
-  | Readonly<[TOption, ...TOption[]]>
-  | [TOption, ...TOption[]];
+export type PicklistOptions<TOption extends string = string> = MaybeReadonly<
+  [TOption, ...TOption[]]
+>;