+++ library/src/schemas/union/union.ts
@@ -3,6 +3,7 @@ import type {
   ErrorMessage,
   Input,
   Issues,
+  MaybeReadonly,
   Output,
 } from '../../types/index.ts';
 import { parseResult, schemaIssue } from '../../utils/index.ts';
@@ -10,7 +11,9 @@ import { parseResult, schemaIssue } from '../../utils/index.ts';
 /**
  * Union options type.
  */
-export type UnionOptions = [BaseSchema, BaseSchema, ...BaseSchema[]];
+export type UnionOptions = MaybeReadonly<
+  [BaseSchema, BaseSchema, ...BaseSchema[]]
+>;
 
 /**
  * Union schema type.